Count Common Factor
This problem is asked in one of the HackerEarth contests.

Problem Statement: Little Robert likes mathematics. Today his teacher has given him two integers and asked him to find out how many integers can divide both the numbers. Would you like to help him in completing his school assignment?

Input Formatting: There are two integers, a and b as input to the program.

Output Formatting: Print the number of common factors of a and b. Both the input value should be in a range of 1 to 10^12.

Example:

Input: 10 15
Output: 2
Explanation: The common factors of 10 and 15 are 1 and 5. So the answer will be 2.